HBX-1992: Reorganize the reverse engineering strategy hierarchy - Rename 'org.hibernate.tool.internal.reveng.SQLTypeMapping' to 'org.hibernate.tool.internal.reveng.strategy.SQLTypeMapping'
